[Add] LoginCheckInterceptor 구현(User, Tag, Group 적용 X), CommentController return 값 UserInfo 추가, return 값이 void 였던 method 전부 응답코드 구현(msg 송신여부는 추후 결정), 수정/삭제 시 해당 UserID값과 Session UserID값 비교하도록 구현, etc, [S03P12A110-99]

+++ backend/src/main/java/io/ssafy/p/i3a110/controller/CommitController.java
@@ -2,7 +2,11 @@ package io.ssafy.p.i3a110.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,13 +15,18 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import io.ssafy.p.i3a110.dto.CommitListDto;
+import io.ssafy.p.i3a110.dto.UserDto;
+import io.ssafy.p.i3a110.interceptor.Auth;
 import io.ssafy.p.i3a110.service.CommitService;
+import io.ssafy.p.i3a110.service.UserService;
 import io.swagger.annotations.ApiOperation;
 
 @RestController
 public class CommitController {
 	@Autowired
 	private CommitService commitService;
+	@Autowired
+	private UserService userService;
 	
 	@GetMapping("/commits/{pid}")
 	@ApiOperation(value = "Post 별 Commit 조회")
@@ -25,15 +34,30 @@ public class CommitController {
 		return commitService.getAllCommitsByPost(pid);
 	}
 	
+	@Auth
 	@PostMapping("/commits")
 	@ApiOperation(value = "포스트에 Commit 추가")
-	public void addCommit(@RequestBody CommitListDto commit) {
+	public Object addCommit(HttpSession session, @RequestBody CommitListDto commit) {
+		String email = (String) session.getAttribute("email");
+		UserDto user = userService.findUserByEmail(email);
+		commit.setUid(user.getId());
+		
 		commitService.addCommit(commit);
+		return new ResponseEntity<>(commit.getId(), HttpStatus.OK);
 	}
 	
+	@Auth
 	@DeleteMapping("/commits/{id}")
 	@ApiOperation(value = "Commit 삭제")
-	public void deleteCommit(@PathVariable String id) {
-		commitService.deleteCommit(id);
+	public Object deleteCommit(HttpSession session, @PathVariable String id) {
+		String email = (String) session.getAttribute("email");
+		UserDto user = userService.findUserByEmail(email);
+		
+		if(user.getId() == commitService.getUidById(id)) {
+			commitService.deleteCommit(id);
+			return new ResponseEntity<>(HttpStatus.OK);
+		}else {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
 	}
 }
